Move SuggestionListKeyboardSelection tests to their own file
