+++ iOS/DuckDuckGoTests/UnifiedToggleInput/DuckAISuggestionsViewControllerTests.swift
@@ -370,45 +370,3 @@ private final class MockSyncPromoManager: SyncPromoManaging {
         dismissedTouchpoints.removeAll()
     }
 }
-
-// MARK: - Keyboard selection math
-
-final class SuggestionListKeyboardSelectionTests: XCTestCase {
-
-    func testWhenNoSelectionAndNextThenFirstItemIsSelected() {
-        XCTAssertEqual(SuggestionListKeyboardSelection.next(after: nil, in: [1, 2, 3]), 1)
-    }
-
-    func testWhenNextFromMiddleThenAdvancesByOne() {
-        XCTAssertEqual(SuggestionListKeyboardSelection.next(after: 2, in: [1, 2, 3]), 3)
-    }
-
-    func testWhenNextAtLastThenClampsToLast() {
-        XCTAssertEqual(SuggestionListKeyboardSelection.next(after: 3, in: [1, 2, 3]), 3)
-    }
-
-    func testWhenNextOnEmptyListThenReturnsNil() {
-        XCTAssertNil(SuggestionListKeyboardSelection.next(after: nil, in: [Int]()))
-    }
-
-    func testWhenNextFromSelectionNotInListThenSelectionIsUnchanged() {
-        XCTAssertEqual(SuggestionListKeyboardSelection.next(after: 99, in: [1, 2, 3]), 99)
-    }
-
-    func testWhenNoSelectionAndPreviousThenStaysUnselected() {
-        XCTAssertNil(SuggestionListKeyboardSelection.previous(before: nil, in: [1, 2, 3]))
-    }
-
-    func testWhenPreviousFromMiddleThenRetreatsByOne() {
-        XCTAssertEqual(SuggestionListKeyboardSelection.previous(before: 2, in: [1, 2, 3]), 1)
-    }
-
-    func testWhenPreviousAtFirstThenClampsToFirst() {
-        XCTAssertEqual(SuggestionListKeyboardSelection.previous(before: 1, in: [1, 2, 3]), 1)
-    }
-
-    func testWhenPreviousFromSelectionNotInListThenSelectionIsUnchanged() {
-        XCTAssertEqual(SuggestionListKeyboardSelection.previous(before: 99, in: [1, 2, 3]), 99)
-    }
-
-}
